% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_appeal_stf.R
\name{classify_appeal_stf}
\alias{classify_appeal_stf}
\title{classify appeals from stf based on decision summary}
\usage{
classify_appeal_stf(x)
}
\arguments{
\item{x}{strings vector}
}
\value{
tibble
}
\description{
classify appeals from stf based on decision summary
}
