% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_details_stf.R
\name{download_details_stf}
\alias{download_details_stf}
\title{Download detailed information based on incidente number}
\usage{
download_details_stf(incidente = NULL, dir = ".")
}
\arguments{
\item{incidente}{incidente number}

\item{dir}{directory}
}
\value{
htmls
}
\description{
Download detailed information based on incidente number
}
