% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stf_collection.R
\name{download_stf_collection}
\alias{download_stf_collection}
\title{Downloads Brazilian Supreme Courts decisions' collection}
\usage{
download_stf_collection(decision_type = NULL, years = NULL,
  dir = ".")
}
\arguments{
\item{decision_type}{either "monocraticas", "presidente" or "colegiadas"}

\item{years}{numeric vector of years}

\item{dir}{directory where to download the files. Defalts to the current dir.}
}
\value{
downloaded files with metadata of all supreme court's decisions
}
\description{
Downloads Brazilian Supreme Courts decisions' collection
}
\examples{
\dontrun{
download_stf_colletion(decision_type = "monocraticas", years = c(2017, 2018), path = ".")
}
}
