% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stf_pdf.R
\name{download_stf_pdf}
\alias{download_stf_pdf}
\title{Downloads pdf text documents based on docket sheet}
\usage{
download_stf_pdf(sheet, path = ".")
}
\arguments{
\item{sheet}{provide tibble read by function read_stf_sheet.}

\item{path}{where to download the pdfs.}
}
\value{
downloaded pdfs with incident number and document id number.
}
\description{
Downloads pdf text documents based on docket sheet
}
\examples{
\dontrun{
download_stf_pdf(sheet=andamento)
}
}
