% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stf_rtf.R
\name{download_stf_rtf}
\alias{download_stf_rtf}
\title{Downloads rtf documents based on docs_url column from follow-up data.}
\usage{
download_stf_rtf(x = NULL, path = ".")
}
\arguments{
\item{x}{docs_url vector gotten from follow-up}

\item{path}{where to download the texts.}
}
\value{
files with rtf texts.
}
\description{
Downloads rtf documents based on docs_url column from follow-up data.
}
\examples{
\dontrun{
download_stf_rtf(x=url)
}
}
