% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{empiricalDistTable}
\alias{empiricalDistTable}
\title{Create a probability table for scores}
\usage{
empiricalDistTable(scores)
}
\arguments{
\item{scores}{\link{numeric} Vector of simulated scores.}
}
\value{
A data frame.
}
\description{
Create a probability table for scores
}
