% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stf_collection.R
\name{get_stf_collection}
\alias{get_stf_collection}
\title{Wraper around functions download_stf_collection and read_stf_collection.}
\usage{
get_stf_collection(decision_type = NULL, years = NULL)
}
\arguments{
\item{decision_type}{either "monocraticas", "presidente" or "colegiadas"}

\item{years}{vector of years}
}
\value{
tibble with metadata of all supreme court's decisions
}
\description{
Wraper around functions download_stf_collection and read_stf_collection.
}
\examples{
\dontrun{
get_stf_colletion(decision_type = "monocraticas", years = c(2017, 2018))
}
}
