% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stf_opinion.R
\name{get_stf_opinion}
\alias{get_stf_opinion}
\title{Get the pdfs with the panel's opinion text  according to the metadata
obtained by the stf_opinion_metadata function.}
\usage{
get_stf_opinion(df)
}
\arguments{
\item{df}{data frame with at least two columns, one with the url,
and the other, named electronico, with a logical vector informing whether the pdf is
a text(TRUE) or an image(FALSE).}
}
\value{
vector with the whole content of the opinions.
}
\description{
Get the pdfs with the panel's opinion text  according to the metadata
obtained by the stf_opinion_metadata function.
}
\keyword{court}
\keyword{opinion,}
\keyword{precedents,}
\keyword{stf,}
\keyword{supreme}
