% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stf_rtf.R
\name{get_stf_rtf}
\alias{get_stf_rtf}
\title{Gets rtf texts from url connection}
\usage{
get_stf_rtf(x)
}
\arguments{
\item{x}{docs_url vector to rtf text.}
}
\value{
a tibble
}
\description{
Gets rtf texts from url connection
}
\details{
This functions is wrapper around download_stf_rtf
and read_stf_rtf. It just returns a tibble with the
text and the text id already parsed, without downloading it
to your local disk.
}
\examples{
\dontrun{
df <- get_stf_rtf()
}
}
