% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stf_thesaurus.R
\name{get_stf_thesaurus}
\alias{get_stf_thesaurus}
\title{Gets STF thesaurus}
\usage{
get_stf_thesaurus(letter = "", term = "")
}
\arguments{
\item{letter}{specify the letter}

\item{term}{specify the word or phrase to be searched}
}
\value{
Tibble with the thesaurus
}
\description{
Gets STF thesaurus
}
\keyword{stj,}
\keyword{thesaurus}
