% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_stf_text.R
\name{join_stf_text}
\alias{join_stf_text}
\title{Joins pdf and rtf texts to docket sheet}
\usage{
join_stf_text(sheet, rtf = NULL, pdf = NULL)
}
\arguments{
\item{sheet}{docket sheet used to download the texts with
functions download_stf_rtf and download_stf_pdf.}

\item{rtf}{tibble with rtf texts}

\item{pdf}{tibble with pdf texts}
}
\value{
new tibble with all texts.
}
\description{
Joins pdf and rtf texts to docket sheet
}
\examples{
\dontrun{
andamento <- join_text(sheet=andamento,rtf=texto_rtf,texto_pdf)
}
}
