% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo.R
\docType{data}
\name{modelo}
\alias{modelo}
\title{modelo}
\format{
charactere:
\describe{
\item{\code{modelo}}{character}
}
}
\source{
\url{https://stf.jus.br/}
}
\usage{
modelo
}
\description{
modelo de json para busca no STF
}
\keyword{datasets}
