% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download-stf.R
\name{parse_stf}
\alias{parse_stf}
\title{Pega andamentos do STF}
\usage{
parse_stf(arqs)
}
\arguments{
\item{arqs}{character vector com nomes dos arquivos html
baixados usando \code{\link{download_stf}}}
}
\description{
Pega andamentos do stf a partir de arquivos já baixados.
}

