% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotScores}
\alias{plotScores}
\title{Plot histogram of simulated scores}
\usage{
plotScores(scores)
}
\arguments{
\item{scores}{Vector of simulated scores}
}
\value{
A ggplot.
}
\description{
Plot histogram of simulated scores
}
