% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{prob}
\alias{prob}
\title{Calculate empirical probability based on simulated scores}
\usage{
prob(x, scores, lower = TRUE)
}
\arguments{
\item{x}{Cutoff for probability of interest.}

\item{scores}{Vector of simulated scores.}

\item{lower}{Logical value. If \code{TRUE}, returns \eqn{P[X \leq x]}, otherwise \eqn{P[X > x]}.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate empirical probability based on simulated scores
}
\examples{
scores = simScores(ns = 40, nw = 120, n = 100)
prob(x = 40, scores = scores)
}
