% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_collection.R
\name{read_stf_collection1}
\alias{read_stf_collection1}
\title{Read STF collection}
\usage{
read_stf_collection1(dir = ".", classes = NULL, years = NULL,
  plan = "sequential")
}
\arguments{
\item{dir}{directory where to find the xlsx files.}

\item{action}{proceding (classe)}

\item{year}{numeric vector of years to read.}
}
\value{
a tibble with all columns read
}
\description{
Read STF collection
}
\examples{
\dontrun{
read_stf_collection(action=NULL,year=NULL)
}
}
