% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_details.R
\name{read_stf_details}
\alias{read_stf_details}
\title{Reads decision details}
\usage{
read_stf_details(files = NULL, path = ".")
}
\arguments{
\item{files}{Files with full path.}

\item{path}{Where to find the htmls}
}
\value{
tibble with meio (public or eletronic), sigilo (secrecy),
numero_unico (unique number), relator_atual (current justice)
}
\description{
Reads decision details
}
\examples{
\dontrun{
detalhes <- read_stf_details(path = ".")
}

}
