% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_docket_sheet.R
\name{read_stf_docket_sheet}
\alias{read_stf_docket_sheet}
\title{Reads Brazilian Supreme Court decision's docket sheet.}
\usage{
read_stf_docket_sheet(files = NULL, path = ".", plan = "sequential")
}
\arguments{
\item{files}{Vector of html files downloaded by download_docket}

\item{path}{Where to find the htmls if files is NULL}

\item{plan}{Defalts to sequential. See \code{future::plan} for all options.}
}
\value{
tibble with follow-up details
}
\description{
Reads Brazilian Supreme Court decision's docket sheet.
}
\examples{
\dontrun{
andamento <- read_stf_docket_sheet(path = ".", plan = "multiprocess")
}

}
