% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_parties.R
\name{read_stf_parties}
\alias{read_stf_parties}
\title{Reads docket parties names.}
\usage{
read_stf_parties(files = NULL, path = ".")
}
\arguments{
\item{files}{Paths to files.}

\item{path}{Where to find the htmls if files NULL.}
}
\value{
Tibble with docket number, party name, and party class.
}
\description{
Reads docket parties names.
}
\examples{
\dontrun{
partes <- read_stf_parties(path = ".")
}

}
