% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_pdf.R
\name{read_stf_pdf}
\alias{read_stf_pdf}
\title{Reads pdf files}
\usage{
read_stf_pdf(file = NULL)
}
\arguments{
\item{file}{of paths to the pdf files to be read.}
}
\value{
a tibble with tree columns: incidente, texto, and doc_id.
}
\description{
Reads pdf files
}
\examples{
\dontrun{
read_stf_pdf(file="")
}
}
