% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{roll}
\alias{roll}
\title{Roll strong and weak dice}
\usage{
roll(ns, nw, ps = 0.4, pw = 0.2)
}
\arguments{
\item{ns}{Number of "strong" dice to roll.}

\item{nw}{Number of "weak" dice to roll.}

\item{ps}{Success probability of "strong" dice. Defaults to \code{0.4}.}

\item{pw}{Success probability of "weak" dice. Defaults to \code{0.2}.}
}
\value{
A vector of length \code{ns + nw} that contains rolls from \code{ns} "strong"
dice and \code{nw} "weak" dice.
}
\description{
Roll strong and weak dice
}
\examples{
roll(ns = 40, nw = 120)
}
