% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{STF shiny server}
\usage{
server(input, output)
}
\arguments{
\item{input}{Input to server.}

\item{output}{Output from server.}
}
\value{
I'm not sure.
}
\description{
STF shiny server
}
