% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle.R
\name{simBattle}
\alias{simBattle}
\title{Simulate a battle}
\usage{
simBattle(a_ns, a_nw, a_multi = 1, d_ns, d_nw, d_multi = 1, n = 1000)
}
\arguments{
\item{a_ns}{Number of attacker "strong" dice to roll for each simulation.}

\item{a_nw}{Number of attacker "weak" dice to roll for each simulation.}

\item{a_multi}{Multiplier that adjusts attacker score from each simulation. Defaults to \code{1}.}

\item{d_ns}{Number of defender "strong" dice to roll for each simulation.}

\item{d_nw}{Number of defender "weak" dice to roll for each simulation.}

\item{d_multi}{Multiplier that adjusts defender score from each simulation. Defaults to \code{1}.}

\item{n}{Number of battles to simulate. Defaults to \code{10000}.}
}
\value{
A numeric vector containing estimated battle outcomes.
}
\description{
Simulate a battle
}
