% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_authenticate.R
\name{stf_authenticate}
\alias{stf_authenticate}
\title{Authenticate to use STF API}
\usage{
stf_authenticate(login = NULL, password = NULL)
}
\arguments{
\item{login}{tax number (CPF) with or without punctiation}

\item{password}{Character password}
}
\value{
TRUE if authenticated
}
\description{
Authenticate to use STF API
}
