% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_cjsg.R
\name{stf_baixar_cjsg}
\alias{stf_baixar_cjsg}
\title{Baixa acórdãos e decisões monocráticas do STF.}
\usage{
stf_baixar_cjsg(
  corpo = NULL,
  livre = "",
  base = c("acordaos", "decisoes"),
  dt_ini = "",
  dt_fim = "",
  dtp_ini = "",
  dtp_fim = "",
  ministros = "",
  classes = "",
  ufs = "",
  tamanho = 10,
  dir = "."
)
}
\arguments{
\item{corpo}{Você pode usar o body(json da sua busca no STF)}

\item{livre}{Termos (use AND ou OR se necessário)}

\item{base}{Acórdãos ou decisões}

\item{dt_ini}{Data início dos julgamentos}

\item{dt_fim}{Data fim dos julgamentos}

\item{dtp_ini}{Data início da publicação}

\item{dtp_fim}{Data final da publicação}

\item{ministros}{Vetor com nome dos ministros. Não funciona ainda.}

\item{classes}{Vetor com as classes processuais. Não funciona ainda.}

\item{ufs}{Vetor com as siglas das unidades federativas. Não funciona ainda.}

\item{tamanho}{Número de decisões por página}

\item{dir}{Diretório}
}
\value{
arquivos em json.
}
\description{
Baixa acórdãos e decisões monocráticas do STF.
}
