% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_details.R
\name{stf_baixar_detalhes}
\alias{stf_baixar_detalhes}
\title{Baixa detalhes do capa com base no número do incidente}
\usage{
stf_baixar_detalhes(incidente = NULL, dir = ".", sono = 1)
}
\arguments{
\item{incidente}{Número do incidente}

\item{dir}{diretório}

\item{sono}{= Sono entre requisições.}
}
\value{
htmls
}
\description{
Baixa detalhes do capa com base no número do incidente
}
