% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_information.R
\name{stf_baixar_informacoes}
\alias{stf_baixar_informacoes}
\title{Baixa html com informação baseada no incidente}
\usage{
stf_baixar_informacoes(incidente, dir = ".", sono = 1)
}
\arguments{
\item{incidente}{Número do incidente}

\item{dir}{Diretório onde armazenar o html}

\item{sono}{= Sono entre requisições.}
}
\value{
htmls
}
\description{
Baixa html com informação baseada no incidente
}
