% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_baixar_partes.R
\name{stf_baixar_partes}
\alias{stf_baixar_partes}
\title{Baixa html das partes com base no incidente}
\usage{
stf_baixar_partes(incidente, dir = ".", sono = 1)
}
\arguments{
\item{incidente}{Número do incidente}

\item{dir}{Diretório onde armazenar os htmls}

\item{sono}{= Sono entre requisições.}
}
\value{
htmls
}
\description{
Baixa html das partes com base no incidente
}
