% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_baixar_pecas.R
\name{stf_baixar_pecas}
\alias{stf_baixar_pecas}
\title{Baixa peças processuais a partir das urls lidas com stj_ler_tabela_pecas.R}
\usage{
stf_baixar_pecas(url_doc, incidente, sequencia, diretorio = ".", sleep = 1)
}
\arguments{
\item{url_doc}{Vetor de urls}

\item{incidente}{Número do incidente}

\item{sequencia}{Ordem das peças}

\item{diretorio}{Onde armazenar}

\item{sleep}{Ao menos um segundo entre requisições.}
}
\value{
pdfs
}
\description{
Baixa peças processuais a partir das urls lidas com stj_ler_tabela_pecas.R
}
