% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_baixar_tabela_pecas.R
\name{stf_baixar_tabela_pecas}
\alias{stf_baixar_tabela_pecas}
\title{Baixa tabela de peças processuais a partir do número do incidente.}
\usage{
stf_baixar_tabela_pecas(incidente, diretorio = ".")
}
\arguments{
\item{incidente}{Vetor de incidentes}

\item{diretorio}{Diretório onde armazenar os arquivos.}
}
\value{
html
}
\description{
Baixa tabela de peças processuais a partir do número do incidente.
}
