% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_classify_appeal.R
\name{stf_classify_appeal}
\alias{stf_classify_appeal}
\title{classifies appeals from stf based on decision summary}
\usage{
stf_classify_appeal(x)
}
\arguments{
\item{x}{strings vector}
}
\value{
tibble
}
\description{
classifies appeals from stf based on decision summary
}
