% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_combinar_pecas.R
\name{stf_combinar_pecas}
\alias{stf_combinar_pecas}
\title{Combina pecas baixados com stf_baixar_pecas}
\usage{
stf_combinar_pecas(
  arquivos = NULL,
  dir_origem = ".",
  dir_destino = NULL,
  nivel = c("incidente", "doc"),
  forcar = TRUE
)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{dir_origem}{Diretório de origem, se não
informar arquivos}

\item{dir_destino}{Diretório destino}

\item{nivel}{Juntar processo inteiro ou por documento?}

\item{forcar}{Padrão para TRUE. Documentos inválidos são excluídos.}
}
\value{
único pdf
}
\description{
Combina pecas baixados com stf_baixar_pecas
}
