% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_cjsg.R
\name{stf_download_cjsg}
\alias{stf_download_cjsg}
\title{Download STF decisions by free search}
\usage{
stf_download_cjsg(
  search = "",
  base = c("acordaos", "decisoes"),
  size = 10,
  dir = "."
)
}
\arguments{
\item{search}{Terms (use AND or OR if needed)}

\item{base}{Acordaos or decisoes}

\item{size}{Number of decisions per page}

\item{dir}{Directory}
}
\value{
json files.
}
\description{
Download STF decisions by free search
}
