% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_details.R
\name{stf_download_details}
\alias{stf_download_details}
\title{Download detailed information based on incidente number}
\usage{
stf_download_details(incidente = NULL, dir = ".")
}
\arguments{
\item{incidente}{incidente number}

\item{dir}{directory}
}
\value{
htmls
}
\description{
Download detailed information based on incidente number
}
