% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_details2.R
\name{stf_download_details2}
\alias{stf_download_details2}
\title{Download details based on classe and docket number}
\usage{
stf_download_details2(class = NULL, docket_number = NULL, dir = ".")
}
\arguments{
\item{class}{This is what in Brazil is called classe processual}

\item{docket_number}{Number of the lawsuit}

\item{dir}{directory under which all subdirectories will
be placed.}
}
\value{
Downloads the html and returns a vector of incident numbers
}
\description{
Download details based on classe and docket number
}
