% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_document_table.R
\name{stf_download_document_table}
\alias{stf_download_document_table}
\title{Download procedural documents (peças) list}
\usage{
stf_download_document_table(incidente, dir = ".")
}
\arguments{
\item{incidente}{Incident number}

\item{dir}{Directory where to put the htmls}
}
\value{
htmls to be read with stf_read_document_table
}
\description{
Download procedural documents (peças) list
}
