% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_documents.R
\name{stf_download_documents}
\alias{stf_download_documents}
\title{Downloads documents from table created
with stf_read_document_table}
\usage{
stf_download_documents(url, dir = ".")
}
\arguments{
\item{url}{Document's url}

\item{dir}{Dir}
}
\value{
pdfs
}
\description{
Downloads documents from table created
with stf_read_document_table
}
