% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_information.R
\name{stf_download_information}
\alias{stf_download_information}
\title{Download html with information based on incidente}
\usage{
stf_download_information(incidente, dir = ".")
}
\arguments{
\item{incidente}{Incidente number}

\item{dir}{Directory where to download the htmls}
}
\value{
htmls
}
\description{
Download html with information based on incidente
}
