% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_lawsuit.R
\name{stf_download_lawsuit}
\alias{stf_download_lawsuit}
\title{Downloads lawsuit}
\usage{
stf_download_lawsuit(incidentes, dir = ".")
}
\arguments{
\item{incidentes}{Vector of incidentes}

\item{dir}{Where do save the json files}
}
\value{
json
}
\description{
Downloads lawsuit
}
