% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_parties.R
\name{stf_download_parties}
\alias{stf_download_parties}
\title{Download html with parties based on incidente}
\usage{
stf_download_parties(incidente, dir = ".")
}
\arguments{
\item{incidente}{Incidente number}

\item{dir}{Directory where to download the htmls}
}
\value{
htmls
}
\description{
Download html with parties based on incidente
}
