% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_sheet.R
\name{stf_download_sheet}
\alias{stf_download_sheet}
\title{Download html with docket follow-up sheet based on incidente}
\usage{
stf_download_sheet(incidente, dir = ".")
}
\arguments{
\item{incidente}{Incidente number}

\item{dir}{Directory where to download the htmls}
}
\value{
htmls
}
\description{
Download html with docket follow-up sheet based on incidente
}
