% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_extract_date.R
\name{stf_extract_date}
\alias{stf_extract_date}
\title{Extract dates from opinions' text}
\usage{
stf_extract_date(x = NULL)
}
\arguments{
\item{x}{Opinion's text}
}
\value{
Vector of dates
}
\description{
Extract dates from opinions' text
}
