% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_get_communication.R
\name{stf_get_communication}
\alias{stf_get_communication}
\title{Gets daily coummunication from STF about
one's lawsuit}
\usage{
stf_get_communication()
}
\value{
tibble
}
\description{
Gets daily coummunication from STF about
one's lawsuit
}
\details{
You need to be authenticated \code{stf_authenticate}
to use this function.
}
