% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_information.R
\name{stf_ler_informacoes}
\alias{stf_ler_informacoes}
\alias{stf_read_information}
\title{Lê dados baixados com stf_baixar_informacoes.R}
\usage{
stf_ler_informacoes(arquivos = NULL, dir = ".")

stf_read_information(arquivos = NULL, dir = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{dir}{Alternativamente informar diretorio.}
}
\value{
tibble
}
\description{
Lê dados baixados com stf_baixar_informacoes.R
}
