% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_ler_tabela_pecas.R
\name{stf_ler_tabela_pecas}
\alias{stf_ler_tabela_pecas}
\title{Lê htmls baixados com stj_baixar_tabela_pecas.}
\usage{
stf_ler_tabela_pecas(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos.}

\item{diretorio}{Alternativamente, informar vetor de arquivos.}
}
\value{
tibble
}
\description{
Lê htmls baixados com stj_baixar_tabela_pecas.
}
