% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_quantidade.R
\name{stf_quantidade}
\alias{stf_quantidade}
\title{Estima quantidade de processos no STF}
\usage{
stf_quantidade(inicio = NULL, fim = NULL)
}
\arguments{
\item{inicio}{Número inicial}

\item{fim}{Número final}
}
\value{
Quantidade distribuída
}
\description{
Estima quantidade de processos no STF
}
