% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_read_cjsg.R
\name{stf_read_cjsg}
\alias{stf_read_cjsg}
\title{Read precedentes downloaded with stf_download_cjsg}
\usage{
stf_read_cjsg(files = NULL, dir = ".")
}
\arguments{
\item{files}{Vector of json files}

\item{dir}{If you don't informe file, provide the dir.}
}
\value{
tibble
}
\description{
Read precedentes downloaded with stf_download_cjsg
}
