% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_read_document_table.R
\name{stf_read_document_table}
\alias{stf_read_document_table}
\title{Read procedutural document list table
donwloaded with stf_download_document_table}
\usage{
stf_read_document_table(files = NULL, path)
}
\arguments{
\item{files}{Vector of htmls files}

\item{path}{Inform if files not provided}
}
\value{
Tibble
}
\description{
Read procedutural document list table
donwloaded with stf_download_document_table
}
