% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_read_parties.R
\name{stf_read_parties}
\alias{stf_read_parties}
\title{Reads docket parties names.}
\usage{
stf_read_parties(files = NULL, path = ".")
}
\arguments{
\item{files}{Paths to files.}

\item{path}{Where to find the htmls if files NULL.}
}
\value{
Tibble with docket number, party name, and party class.
}
\description{
Reads docket parties names.
}
\examples{
\dontrun{
partes <- stf_read_parties(path = ".")
}

}
