% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_read_pdf.R
\name{stf_read_pdf}
\alias{stf_read_pdf}
\title{Reads pdf files}
\usage{
stf_read_pdf(files = NULL, path = ".")
}
\arguments{
\item{files}{Vector of pdf files  to be read.}

\item{path}{Path to pdf files if files are not informed}
}
\value{
a tibble with tree columns: incidente, texto, and doc_id.
}
\description{
Reads pdf files
}
\examples{
\dontrun{
pdf_read_stf(files = "")
}
}
