% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_stf_help.R
\name{view_stf_help}
\alias{view_stf_help}
\title{Displays STF's help page on browser.}
\usage{
view_stf_help()
}
\value{
This function has only a side effect, it doesn't return any object or file.
}
\description{
Displays STF's help page on browser.
}
\examples{
\dontrun{
view_stf_help()
}
}
\keyword{help}
\keyword{stf,}
