% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_stf_thesaurus.R
\name{view_stf_thesaurus}
\alias{view_stf_thesaurus}
\title{Function stf_thesaurus}
\usage{
view_stf_thesaurus()
}
\value{
This function has only side effect, it doesn't return anything.
}
\description{
This function displays STF's thesaurus.
}
\examples{
\dontrun{
view_stf_thesaurus()
}
}
\keyword{stf,}
\keyword{thesaurus}
