% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stf_dockets.R
\name{download_stf_dockets}
\alias{download_stf_dockets}
\title{Downloads Brazilian Supreme Court lawsuits based on type of action and docket number.}
\usage{
download_stf_dockets(action = NULL, docket_number = NULL)
}
\arguments{
\item{action}{This is what in Brazil is called classe processual}

\item{docket_number}{Number of the lawsuit}
}
\value{
htmls in nine folders corresponding to each one of the stf webpage tabs.
}
\description{
Downloads Brazilian Supreme Court lawsuits based on type of action and docket number.
}
\examples{
\dontrun{
download_stf_dockets(action="HC",docket_number="4050")
}
}
