% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stf_precedent_sheet.R
\name{get_stf_precedent_sheet}
\alias{get_stf_precedent_sheet}
\title{Title Downloads STF precedent sheet}
\usage{
get_stf_precedent_sheet(url)
}
\arguments{
\item{url}{urls from precedent sheet.}
}
\value{
tibble with the docket sheet from precedent search
}
\description{
Title Downloads STF precedent sheet
}
\keyword{andamento}
\keyword{jurisprudência,}
\keyword{processual}
\keyword{stf,}
