% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_collection.R
\name{read_stf_collection}
\alias{read_stf_collection}
\title{Read STF collection}
\usage{
read_stf_collection(dir = ".", classes = NULL, years = NULL,
  plan = "sequential")
}
\arguments{
\item{dir}{directory where to find the xlsx files.}

\item{classes}{proceding (classe)}

\item{years}{numeric vector of years to read.}

\item{plan}{default to "sequential". Check future::plan for datails.}
}
\value{
a tibble with all columns read
}
\description{
Read STF collection
}
\examples{
\dontrun{
read_stf_collection(action=NULL,year=NULL)
}
}
