% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_collection.R
\name{read_stf_collection}
\alias{read_stf_collection}
\title{Read STF collection}
\usage{
read_stf_collection(files = NULL, dir = ".", classes = NULL, years = NULL)
}
\arguments{
\item{files}{Vector of xlsx files}

\item{dir}{directory where to find the xlsx files if files is NULL}

\item{classes}{proceding (classe)}

\item{years}{numeric vector of years to read.}
}
\value{
a tibble with all columns read
}
\description{
Read STF collection
}
\examples{
\dontrun{
read_stf_collection(classes = NULL, years = 2017:2018)
}
}
