% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_details.R
\name{read_stf_details}
\alias{read_stf_details}
\title{Read decision details}
\usage{
read_stf_details(path = ".", plan = "sequential")
}
\arguments{
\item{path}{where to find the htmls}

\item{plan}{default to "sequential", see \code{future::plan},
for all options.}
}
\value{
tibble with meio (public or eletronic), sigilo (secrecy),
numero_unico (unique number), relator_atual (current justice)
}
\description{
Read decision details
}
\examples{
\dontrun{
detalhes <- read_stf_details(path=".",plan="multiprocess")
}

}
