% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_docket_sheet.R
\name{read_stf_docket_sheet}
\alias{read_stf_docket_sheet}
\title{Reads Brazilian Supreme Court decision's docket sheet.}
\usage{
read_stf_docket_sheet(path = ".", plan = "sequential")
}
\arguments{
\item{path}{where to find the htmls downloaded by download_docket}

\item{plan}{defalts to sequential. See \code{future::plan} for all options.}
}
\value{
tibble with follow-up details
}
\description{
Reads Brazilian Supreme Court decision's docket sheet.
}
\examples{
\dontrun{
andamento <- read_stf_docket_sheet(path=".",plan="multiprocess")
}

}
