% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_information.R
\name{read_stf_information}
\alias{read_stf_information}
\alias{stf_read_information}
\title{Reads data from information tab.}
\usage{
read_stf_information(files = NULL, path = ".")

stf_read_information(files = NULL, path = ".")
}
\arguments{
\item{files}{Vector of html files.}

\item{path}{where the htmls are if files is NULL.}
}
\value{
a tibble with data from the information tab.
}
\description{
Reads data from information tab.
}
\examples{
\dontrun{
informacao <- read_stf_information("html", plan = "multiprocess")
}
}
