% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stf_parties.R
\name{read_stf_parties}
\alias{read_stf_parties}
\title{Reads docket parties names.}
\usage{
read_stf_parties(path = ".", plan = "sequential")
}
\arguments{
\item{path}{where to find the htmls.}

\item{plan}{default to "sequential". See \code{future::plan}
for all options.}
}
\value{
Tibble with docket number, party name, and party class.
}
\description{
Reads docket parties names.
}
\examples{
\dontrun{
partes<- read_stf_parties(path=".",plan="multiprocess")
}

}
