% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf-package.R
\docType{package}
\name{stf-package}
\alias{stf}
\alias{stf-package}
\title{stf: Functions to collect and manipulate Brazilian's Supreme Court decisions}
\description{
Retrieves data from Brazilian Supreme Court decisions.
}
\author{
\strong{Maintainer}: Jose de Jesus Filho \email{jjesus@abj.org.br}

}
\keyword{internal}
