% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf-package.R
\docType{package}
\name{stf-package}
\alias{stf}
\alias{stf-package}
\title{stf: Provides functions to collect and manipulate Brazilian's Supreme
    Court decisions}
\description{
Retrieves data from Brazilian Supreme Court
    decisions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jjesusfilho.github.io/stf/}
}

}
\author{
\strong{Maintainer}: Jose de Jesus Filho \email{jose@consudata.com.br}

Other contributors:
\itemize{
  \item Lívia Pitelli Zamarian Houaiss \email{liviapzamarian@hotmail.com} [contributor]
}

}
\keyword{internal}
