% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\docType{package}
\name{stf-package}
\alias{stf}
\alias{stf-package}
\title{\code{stf} package}
\description{
Downloads and organizes Brazilian Supreme Court's Decisions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jjesusfilho.github.io/stf/}
}

}
\author{
\strong{Maintainer}: Jose de Jesus Filho \email{jose@consudata.com.br}

Other contributors:
\itemize{
  \item Lívia Pitelli Zamarian Houaiss \email{liviapzamarian@hotmail.com} [contributor]
}

}
