% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_download_parties.R
\name{stf_baixar_partes}
\alias{stf_baixar_partes}
\title{Baixa html das partes com base no incidente}
\usage{
stf_baixar_partes(incidente, dir = ".", sono = 1)
}
\arguments{
\item{incidente}{Número do incidente}

\item{sono}{= Sono entre requisições.}

\item{dirs}{Diretório onde armazenar os htmls}
}
\value{
htmls
}
\description{
Baixa html das partes com base no incidente
}
